+++ src/Controller/BaseController.php
@@ -25,10 +25,11 @@ use Template_Helper;
  *
  * @property-read Helper\AssignHelper $assign
  * @property-read Helper\AttachHelper $attach
- * @property-read Helper\HtmlHelper $html
- * @property-read Helper\PlotHelper $plot
- * @property-read Helper\MessagesHelper $messages
  * @property-read Helper\CsrfHelper $csrf
+ * @property-read Helper\HtmlHelper $html
+ * @property-read Helper\MessagesHelper $messages
+ * @property-read Helper\PlotHelper $plot
+ * @property-read Helper\RepositoryHelper $repository
  */
 abstract class BaseController
 {